dialogbert_requirements.txt
absl-py==0.11.0
alabaster==0.7.12
allennlp==0.9.0
attrs==20.2.0
Babel==2.8.0
blis==0.2.4
boto3==1.16.7
botocore==1.19.7
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
checksumdir==1.2.0
click==7.1.2
conllu==1.3.1
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
dataclasses==0.7
docutils==0.16
editdistance==0.5.3
embeddings==0.0.8
filelock==3.0.12
flaky==3.7.0
Flask==1.1.2
Flask-Cors==3.0.9
ftfy==5.8
future==0.18.2
fuzzywuzzy==0.18.0
gevent==20.9.0
google-auth==1.22.1
google-auth-oauthlib==0.4.2
greenlet==0.4.17
grpcio==1.33.2
h5py==2.10.0
idna==2.10
imagesize==1.2.0
importlib-metadata==2.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonnet==0.16.0
jsonpatch==1.26
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==2.6.0
kiwisolver==1.3.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
murmurhash==1.0.2
nltk==3.5
numpy==1.19.3
numpydoc==1.1.0
oauthlib==3.1.0
overrides==3.1.0
packaging==20.4
parsimonious==0.8.1
Pillow==8.0.1
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.2
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
quadprog==0.1.7
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
responses==0.12.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.20.3
scipy==1.5.3
sentencepiece==0.1.94
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.0.0
spacy==2.1.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
srsly==1.0.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==1.7
thinc==7.0.8
tokenizers==0.7.0
toml==0.10.1
torch==1.5.1
torchfile==0.1.0
tornado==6.0.4
tqdm==4.51.0
transformers==2.11.0
Unidecode==1.1.1
urllib3==1.25.11
visdom==0.1.8.9
wasabi==0.8.0
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
word2number==1.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2
